requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
gunicorn==20.0.0
requests==2.22.0
requests-oauthlib==1.3.0
oauthlib==3.1.0
urllib3==1.25.7
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
httplib2==0.14.0
Cerberus==1.3.2